Refaktorointi IV, app.js
